Fix logic error and faiss version
The code relied on an error to be thrown in order to work correctly.
Changed this code to an appropriate if-else statement, and added a
'finally' clause to the 'try' statement.

A windows specific(?) build of faiss was used in the environment.yml. This
makes the conda environment incompatible with different operation systems.
Removed the extra strict version specification.

Added the download directory to the .gitignore file